66 - React JS - урок redux-thunk в деталях (практика)

+++ src/api/api.js
@@ -3,25 +3,27 @@ import axios from "axios";
 const socialNetworkApi = axios.create({
   baseURL: "https://social-network.samuraijs.com/api/1.0",
   withCredentials: true,
+  headers: {
+    "API-KEY": "9cebe972-9f60-4460-94e5-2d7054eec31d",
+  }
 });
 
-export const getUsers = (currentPage = 1, pageSize = 10) =>
-  socialNetworkApi
-    .get("/users", {
-      params: {
-        page: currentPage,
-        count: pageSize,
-      },
-    })
-    .then((response) => response.data);
+export const SocialNetworkAPI = {
+  getUsers: (currentPage = 1, pageSize = 10) =>
+    socialNetworkApi
+      .get("/users", {
+        params: {
+          page: currentPage,
+          count: pageSize,
+        },
+      })
+      .then((response) => response.data),
 
-export const authMe = () => socialNetworkApi.get("/authMe");
+  authMe: () => socialNetworkApi.get("/auth/me"),
 
-export const getUserProfile = (userId) =>
-  socialNetworkApi.get(`/profile/${userId}`);
+  getUserProfile: (userId) => socialNetworkApi.get(`/profile/${userId}`),
 
-export const followUser = (userId) =>
-  socialNetworkApi.post(`/follow/${userId}`, null);
+  followUser: (userId) => socialNetworkApi.post(`/follow/${userId}`, null),
 
-export const unfollowUser = (userId) =>
-  socialNetworkApi.delete(`/follow/${userId}`);
+  unfollowUser: (userId) => socialNetworkApi.delete(`/follow/${userId}`),
+};
